requirements.txt
streamlit
numpy
Pillow
tensorflow
opencv-python-headless
opencv-python
clerk